Delete static repository
