+++ Keda.Cosmosdb.Scaler/src/Repository/ICosmosDBEstimator.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.Azure.Cosmos;
+using System.Threading.Tasks;
+using Keda.CosmosDB.Scaler.Services;
+
+namespace Keda.CosmosDB.Scaler.Repository
+{
+    public interface ICosmosDBEstimator
+    {
+        Task<long> GetEstimatedWork(CosmosDBTrigger trigger);
+    }
+}
