--- Keda.Cosmosdb.Scaler/src/Repository/ICosmosDBRepository.cs
@@ -1,10 +1,0 @@
-﻿using Microsoft.Azure.Cosmos;
-using System.Threading.Tasks;
-
-namespace Keda.CosmosDB.Scaler.Repository
-{
-    public interface ICosmosDBRepository
-    {
-        Task<long> GetEstimatedWork(ChangeFeedEstimator estimator);
-    }
-}
